♻️refactor: InsightHeader/InsightItemを共通ウィジェットに統一（カテゴリ間での重複コードを解消するため）
各クイズカテゴリのリザルト画面に点在していた _InsightHeader / _InsightItem の
重複定義を quiz_core パッケージの QuizInsightHeader / QuizInsightItem として一元化。
同時に shopping の ShoppingInsightItem（黄色背景・ボーダー付き）も削除し、
全カテゴリで共通の見た目に統一した。
