+++ packages/quiz_core/lib/src/widgets/quiz_insight_widgets.dart
@@ -1,0 +1,90 @@
+import 'package:flutter/material.dart';
+
+/// クイズリザルトの「解説」セクションで使うヘッダー。
+/// 電球アイコン＋タイトル＋サブタイトルの3要素で構成する。
+class QuizInsightHeader extends StatelessWidget {
+  const QuizInsightHeader({
+    super.key,
+    required this.title,
+    required this.subtitle,
+  });
+
+  final String title;
+  final String subtitle;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Row(
+          children: [
+            const Icon(Icons.lightbulb, color: Color(0xFFFFD814), size: 20),
+            const SizedBox(width: 6),
+            Expanded(
+              child: Text(
+                title,
+                style: Theme.of(context).textTheme.titleSmall?.copyWith(
+                      fontWeight: FontWeight.bold,
+                    ),
+              ),
+            ),
+          ],
+        ),
+        const SizedBox(height: 4),
+        Text(
+          subtitle,
+          style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                color: Theme.of(context).colorScheme.onSurfaceVariant,
+              ),
+        ),
+      ],
+    );
+  }
+}
+
+/// クイズリザルトの「解説」セクションで使う個別アイテム。
+/// 絵文字＋タイトル＋説明文の3要素で構成する。
+class QuizInsightItem extends StatelessWidget {
+  const QuizInsightItem({
+    super.key,
+    required this.emoji,
+    required this.title,
+    required this.desc,
+  });
+
+  final String emoji;
+  final String title;
+  final String desc;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(emoji, style: const TextStyle(fontSize: 18)),
+        const SizedBox(width: 8),
+        Expanded(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                title,
+                style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                      fontWeight: FontWeight.bold,
+                    ),
+              ),
+              const SizedBox(height: 2),
+              Text(
+                desc,
+                style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                      color: Theme.of(context).colorScheme.onSurfaceVariant,
+                    ),
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
