+++ packages/quizzes/alarm/lib/src/presentation/quiz3_disable_snooze/disable_snooze_quiz_screen.dart
@@ -118,106 +118,24 @@ class _DisableSnoozeInsight extends StatelessWidget {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        _InsightHeader(title: insight.title, subtitle: insight.subtitle),
+        QuizInsightHeader(title: insight.title, subtitle: insight.subtitle),
         const SizedBox(height: 12),
-        _InsightItem(
+        QuizInsightItem(
           emoji: '🔘',
           title: insight.toggleTitle,
           desc: insight.toggleDesc,
         ),
         const SizedBox(height: 10),
-        _InsightItem(
+        QuizInsightItem(
           emoji: '↔️',
           title: insight.vs_arrow_title,
           desc: insight.vs_arrow_desc,
         ),
         const SizedBox(height: 10),
-        _InsightItem(
+        QuizInsightItem(
           emoji: '🎨',
           title: insight.colorTitle,
           desc: insight.colorDesc,
-        ),
-      ],
-    );
-  }
-}
-
-class _InsightHeader extends StatelessWidget {
-  const _InsightHeader({required this.title, required this.subtitle});
-
-  final String title;
-  final String subtitle;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            const Icon(Icons.lightbulb, color: Color(0xFFFFD814), size: 20),
-            const SizedBox(width: 6),
-            Expanded(
-              child: Text(
-                title,
-                style: Theme.of(context).textTheme.titleSmall?.copyWith(
-                      fontWeight: FontWeight.bold,
-                    ),
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 4),
-        Text(
-          subtitle,
-          style: Theme.of(context)
-              .textTheme
-              .bodySmall
-              ?.copyWith(color: Theme.of(context).colorScheme.onSurfaceVariant),
-        ),
-      ],
-    );
-  }
-}
-
-class _InsightItem extends StatelessWidget {
-  const _InsightItem({
-    required this.emoji,
-    required this.title,
-    required this.desc,
-  });
-
-  final String emoji;
-  final String title;
-  final String desc;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(emoji, style: const TextStyle(fontSize: 18)),
-        const SizedBox(width: 8),
-        Expanded(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                title,
-                style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                      fontWeight: FontWeight.bold,
-                    ),
-              ),
-              const SizedBox(height: 2),
-              Text(
-                desc,
-                style: Theme.of(context)
-                    .textTheme
-                    .bodySmall
-                    ?.copyWith(color: Theme.of(context).colorScheme.onSurfaceVariant),
-              ),
-            ],
-          ),
         ),
       ],
     );
