+++ packages/quizzes/chat/lib/src/presentation/quiz4_create_group/create_group_quiz_screen.dart
@@ -499,78 +499,13 @@ class _CreateGroupInsight extends StatelessWidget {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        Row(
-          children: [
-            const Icon(Icons.lightbulb, color: Color(0xFFFFD814), size: 20),
-            const SizedBox(width: 6),
-            Expanded(
-              child: Text(
-                insight.title,
-                style: Theme.of(context).textTheme.titleSmall?.copyWith(
-                      fontWeight: FontWeight.bold,
-                    ),
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 4),
-        Text(
-          insight.subtitle,
-          style: Theme.of(context)
-              .textTheme
-              .bodySmall
-              ?.copyWith(color: Theme.of(context).extension<ChatAppTheme>()!.subTextColor),
-        ),
+        QuizInsightHeader(title: insight.title, subtitle: insight.subtitle),
         const SizedBox(height: 12),
-        _InsightItem(emoji: '➕', title: insight.plusTitle, desc: insight.plusDesc),
+        QuizInsightItem(emoji: '➕', title: insight.plusTitle, desc: insight.plusDesc),
         const SizedBox(height: 10),
-        _InsightItem(emoji: '🧭', title: insight.wizardTitle, desc: insight.wizardDesc),
+        QuizInsightItem(emoji: '🧭', title: insight.wizardTitle, desc: insight.wizardDesc),
         const SizedBox(height: 10),
-        _InsightItem(emoji: '☑️', title: insight.checkboxTitle, desc: insight.checkboxDesc),
-      ],
-    );
-  }
-}
-
-class _InsightItem extends StatelessWidget {
-  const _InsightItem({
-    required this.emoji,
-    required this.title,
-    required this.desc,
-  });
-
-  final String emoji;
-  final String title;
-  final String desc;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(emoji, style: const TextStyle(fontSize: 18)),
-        const SizedBox(width: 8),
-        Expanded(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                title,
-                style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                      fontWeight: FontWeight.bold,
-                    ),
-              ),
-              const SizedBox(height: 2),
-              Text(
-                desc,
-                style: Theme.of(context)
-                    .textTheme
-                    .bodySmall
-                    ?.copyWith(color: Theme.of(context).extension<ChatAppTheme>()!.subTextColor),
-              ),
-            ],
-          ),
-        ),
+        QuizInsightItem(emoji: '☑️', title: insight.checkboxTitle, desc: insight.checkboxDesc),
       ],
     );
   }
