+++ packages/quizzes/payment/lib/src/presentation/quiz2_reveal_balance/reveal_balance_quiz_screen.dart
@@ -94,109 +94,24 @@ class _RevealBalanceInsight extends StatelessWidget {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        _InsightHeader(title: insight.title, subtitle: insight.subtitle),
+        QuizInsightHeader(title: insight.title, subtitle: insight.subtitle),
         const SizedBox(height: 12),
-        _InsightItem(
+        QuizInsightItem(
           emoji: '👁️',
           title: insight.eyeTitle,
           desc: insight.eyeDesc,
         ),
         const SizedBox(height: 10),
-        _InsightItem(
+        QuizInsightItem(
           emoji: '***',
           title: insight.maskTitle,
           desc: insight.maskDesc,
         ),
         const SizedBox(height: 10),
-        _InsightItem(
+        QuizInsightItem(
           emoji: '🔒',
           title: insight.privacyTitle,
           desc: insight.privacyDesc,
-        ),
-      ],
-    );
-  }
-}
-
-class _InsightHeader extends StatelessWidget {
-  const _InsightHeader({required this.title, required this.subtitle});
-
-  final String title;
-  final String subtitle;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            const Icon(Icons.lightbulb, color: Color(0xFFFFD814), size: 20),
-            const SizedBox(width: 6),
-            Expanded(
-              child: Text(
-                title,
-                style: Theme.of(context).textTheme.titleSmall?.copyWith(
-                      fontWeight: FontWeight.bold,
-                    ),
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 4),
-        Text(
-          subtitle,
-          style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                color: Theme.of(context)
-                    .extension<PaymentAppTheme>()!
-                    .secondaryTextColor,
-              ),
-        ),
-      ],
-    );
-  }
-}
-
-class _InsightItem extends StatelessWidget {
-  const _InsightItem({
-    required this.emoji,
-    required this.title,
-    required this.desc,
-  });
-
-  final String emoji;
-  final String title;
-  final String desc;
-
-  @override
-  Widget build(BuildContext context) {
-    final secondaryTextColor =
-        Theme.of(context).extension<PaymentAppTheme>()!.secondaryTextColor;
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(emoji, style: const TextStyle(fontSize: 18)),
-        const SizedBox(width: 8),
-        Expanded(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                title,
-                style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                      fontWeight: FontWeight.bold,
-                    ),
-              ),
-              const SizedBox(height: 2),
-              Text(
-                desc,
-                style: Theme.of(context)
-                    .textTheme
-                    .bodySmall
-                    ?.copyWith(color: secondaryTextColor),
-              ),
-            ],
-          ),
         ),
       ],
     );
