+++ packages/quizzes/shopping/lib/src/presentation/shopping_app.dart
@@ -1750,59 +1750,3 @@ class AmazonCartSheet extends StatelessWidget {
   }
 }
 
-// ─── 共有インサイトアイテム ────────────────────────────────────────────────
-
-/// リザルト画面の UX 解説セクションで使う個別アイテム Widget。
-/// 水クイズ・再注文クイズで共有する。
-class ShoppingInsightItem extends StatelessWidget {
-  const ShoppingInsightItem({
-    super.key,
-    required this.emoji,
-    required this.title,
-    required this.desc,
-  });
-
-  final String emoji;
-  final String title;
-  final String desc;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(10),
-      decoration: BoxDecoration(
-        color: const Color(0xFFFFFBEB),
-        borderRadius: BorderRadius.circular(8),
-        border: Border.all(color: const Color(0xFFFFE082)),
-      ),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(emoji, style: const TextStyle(fontSize: 20)),
-          const SizedBox(width: 10),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  title,
-                  style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                        fontWeight: FontWeight.bold,
-                      ),
-                ),
-                const SizedBox(height: 2),
-                Text(
-                  desc,
-                  style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                        color: Theme.of(context).extension<ShoppingAppTheme>()!.primaryTextColor,
-                        height: 1.4,
-                      ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
